ECOS_cleanup {ECOSolveR}	R Documentation

Clean up an ECOS workspace

Description

Frees the C-level ECOS workspace. After this call the workspace external pointer is invalidated and cannot be used for further solves or updates. It is safe to call this more than once.

Usage

ECOS_cleanup(workspace)

Arguments

workspace
an external pointer of class "ecos_workspace" as returned by ECOS_setup.

Value

NULL, invisibly.

See Also

ECOS_setup, ECOS_solve, ECOS_update

[Package ECOSolveR version 0.6.1 Index]