make_csc_matrix {ECOSolveR}	R Documentation

Convert a plain matrix or simple triplet form matrix to a [Matrix::dgCMatrix-class] (implicit) form

Description

Convert a plain matrix or simple triplet form matrix to a [Matrix::dgCMatrix-class] (implicit) form

Usage

make_csc_matrix(x)

Arguments

x
a matrix or a simple triplet form matrix

Value

a list of row pointer, column pointer, and values corresponding to a [Matrix::dgCMatrix-class] object

[Package ECOSolveR version 0.6.1 Index]